+++ src/Symfony/Bridge/Twig/Extension/HttpKernelExtension.php
@@ -30,6 +30,7 @@ final class HttpKernelExtension extends AbstractExtension
         return [
             new TwigFunction('render', [HttpKernelRuntime::class, 'renderFragment'], ['is_safe' => ['html']]),
             new TwigFunction('render_*', [HttpKernelRuntime::class, 'renderFragmentStrategy'], ['is_safe' => ['html']]),
+            new TwigFunction('fragment_uri', [HttpKernelRuntime::class, 'generateFragmentUri']),
             new TwigFunction('controller', static::class.'::controller'),
         ];
     }