+++ src/Symfony/Bundle/FrameworkBundle/Tests/Functional/FragmentTest.php
@@ -44,4 +44,12 @@ TXT
             [true],
         ];
     }
+
+    public function testGenerateFragmentUri()
+    {
+        $client = self::createClient(['test_case' => 'Fragment', 'root_config' => 'config.yml', 'debug' => true]);
+        $client->request('GET', '/fragment_uri');
+
+        $this->assertSame('/_fragment?_hash=CCRGN2D%2FoAJbeGz%2F%2FdoH3bNSPwLCrmwC1zAYCGIKJ0E%3D&_path=_format%3Dhtml%26_locale%3Den%26_controller%3DSymfony%255CBundle%255CFrameworkBundle%255CTests%255CFunctional%255CBundle%255CTestBundle%255CController%255CFragmentController%253A%253AindexAction', $client->getResponse()->getContent());
+    }
 }