+++ src/Symfony/Component/HttpKernel/Fragment/RoutableFragmentRenderer.php
@@ -22,7 +22,10 @@ use Symfony\Component\HttpKernel\EventListener\FragmentListener;
  */
 abstract class RoutableFragmentRenderer implements FragmentRendererInterface
 {
-    private $fragmentPath = '/_fragment';
+    /**
+     * @internal
+     */
+    protected $fragmentPath = '/_fragment';
 
     /**
      * Sets the fragment path that triggers the fragment listener.
@@ -44,43 +47,6 @@ abstract class RoutableFragmentRenderer implements FragmentRendererInterface
      */
     protected function generateFragmentUri(ControllerReference $reference, Request $request, bool $absolute = false, bool $strict = true)
     {
-        if ($strict) {
-            $this->checkNonScalar($reference->attributes);
-        }
-
-        // We need to forward the current _format and _locale values as we don't have
-        // a proper routing pattern to do the job for us.
-        // This makes things inconsistent if you switch from rendering a controller
-        // to rendering a route if the route pattern does not contain the special
-        // _format and _locale placeholders.
-        if (!isset($reference->attributes['_format'])) {
-            $reference->attributes['_format'] = $request->getRequestFormat();
-        }
-        if (!isset($reference->attributes['_locale'])) {
-            $reference->attributes['_locale'] = $request->getLocale();
-        }
-
-        $reference->attributes['_controller'] = $reference->controller;
-
-        $reference->query['_path'] = http_build_query($reference->attributes, '', '&');
-
-        $path = $this->fragmentPath.'?'.http_build_query($reference->query, '', '&');
-
-        if ($absolute) {
-            return $request->getUriForPath($path);
-        }
-
-        return $request->getBaseUrl().$path;
-    }
-
-    private function checkNonScalar(array $values)
-    {
-        foreach ($values as $key => $value) {
-            if (\is_array($value)) {
-                $this->checkNonScalar($value);
-            } elseif (!is_scalar($value) && null !== $value) {
-                throw new \LogicException(sprintf('Controller attributes cannot contain non-scalar/non-null values (value for key "%s" is not a scalar or null).', $key));
-            }
-        }
+        return (new FragmentUriGenerator($this->fragmentPath))->generate($reference, $request, $absolute, $strict, false);
     }
 }